refs #937 Move organization switcher logic to its own component
