+++ src/components/Dashboard/Menu/Options.tsx
@@ -1,17 +1,12 @@
 import { Box, Collapse } from '@chakra-ui/react'
-import { OrganizationName } from '@vocdoni/chakra-components'
-import { Select } from 'chakra-react-select'
-import { useEffect, useMemo, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { HiSquares2X2 } from 'react-icons/hi2'
 import { IoIosSettings } from 'react-icons/io'
 import { generatePath, matchPath, useLocation } from 'react-router-dom'
-import { useQueryClient } from 'wagmi'
-import { useAuth } from '~components/Auth/useAuth'
-import { LocalStorageKeys } from '~components/Auth/useAuthProvider'
-import { Organization, useProfile } from '~src/queries/account'
 import { Routes } from '~src/router/routes'
 import { DashboardMenuItem } from './Item'
+import { OrganizationSwitcher } from './OrganizationSwitcher'
 
 type MenuItem = {
   label: string
@@ -24,45 +19,6 @@ export const DashboardMenuOptions = () => {
   const { t } = useTranslation()
   const location = useLocation()
   const [openSection, setOpenSection] = useState<string | null>(null)
-  const { data: profile } = useProfile()
-  const [selectedOrg, setSelectedOrg] = useState<string | null>(localStorage.getItem(LocalStorageKeys.SignerAddress))
-  const { signerRefresh } = useAuth()
-  const client = useQueryClient()
-
-  const organizations = useMemo(() => {
-    if (!profile?.organizations) return []
-    return profile.organizations.map((org) => ({
-      value: org.organization.address,
-      label: org.organization.address,
-      organization: org.organization,
-    }))
-  }, [profile])
-
-  // Set first organization as default if none selected
-  useEffect(() => {
-    if (organizations.length && !selectedOrg) {
-      const firstOrgAddress = organizations[0].value
-      setSelectedOrg(firstOrgAddress)
-      localStorage.setItem(LocalStorageKeys.SignerAddress, firstOrgAddress)
-    }
-  }, [organizations, selectedOrg])
-
-  type SelectOption = {
-    value: string
-    label: string
-    organization: Organization
-  }
-
-  const handleOrgChange = async (option: SelectOption | null) => {
-    if (!option) return
-    setSelectedOrg(option.value)
-    localStorage.setItem(LocalStorageKeys.SignerAddress, option.value)
-    // clear all query client query cache
-    client.clear()
-    // refresh signer
-    await signerRefresh()
-  }
-
   const menuItems: MenuItem[] = [
     // {
     //   label: t('organization.dashboard'),
@@ -125,24 +81,7 @@ export const DashboardMenuOptions = () => {
 
   return (
     <Box>
-      {organizations.length > 1 ? (
-        <Box mb={2} px={3.5}>
-          <Select
-            value={organizations.find((org) => org.value === selectedOrg)}
-            onChange={handleOrgChange}
-            options={organizations}
-            size='sm'
-            chakraStyles={{
-              container: (provided) => ({
-                ...provided,
-                width: '100%',
-              }),
-            }}
-          />
-        </Box>
-      ) : (
-        <OrganizationName mb={2} px={3.5} />
-      )}
+      <OrganizationSwitcher />
 
       {menuItems.map((item, index) => (
         <Box key={index}>
